(1/4) decompiler: Add 'using namespace std' to all '.cc'
This commit _should_ compile with all Ghidra code.

It is less controversial (but still not ideal) to add 'using namespace'
to C++ source files. Source files are isolated and will not affect any
code other than what's in the source file. Eventually, it would be best
to refactor to either use a using-declaration statement or to attach
'std::' to all the symbols. Clangd refactors exist to do this (but they
don't work on header files).

I modified the *.y and *.l files, but I did not regenerate because it
would cause more diffs than necessary, but if you did regenerate, it
would generate the intended modifications made to the generated files.